refactor(abstraction): update person type abstraction
fix #179
