+++ src/HospitalRun.tsx
@@ -7,9 +7,9 @@ import NewAppointment from 'scheduling/appointments/new/NewAppointment'
 import Sidebar from './components/Sidebar'
 import Permissions from './model/Permissions'
 import Dashboard from './dashboard/Dashboard'
-import Patients from './patients/list/Patients'
-import NewPatient from './patients/new/NewPatient'
-import ViewPatient from './patients/view/ViewPatient'
+import Patients from './persons/patients/list/Patients'
+import NewPatient from './persons/patients/new/NewPatient'
+import ViewPatient from './persons/patients/view/ViewPatient'
 import { RootState } from './store'
 import Navbar from './components/Navbar'
 import PrivateRoute from './components/PrivateRoute'
