+++ src/__tests__/persons/patients/related-persons/NewRelatedPersonModal.test.tsx
@@ -94,7 +94,7 @@ describe('New Related Person Modal', () => {
       })
       expect(onSaveSpy).toHaveBeenCalledTimes(1)
       expect(onSaveSpy).toHaveBeenCalledWith({
-        relatedPatientId: '123',
+        relatedPersonId: '123',
         type: 'relationship',
       })
     })
