+++ src/model/RelatedPerson.ts
@@ -1,6 +1,6 @@
 import Person from './Person'
 
 export default interface RelatedPerson extends Person {
-  relatedPatientId: string
+  relatedPersonId: string
   relationType: string
 }
