+++ src/persons/patients/new/NewPatient.tsx
@@ -3,8 +3,8 @@ import { useHistory } from 'react-router'
 import { useTranslation } from 'react-i18next'
 import { useDispatch } from 'react-redux'
 import NewPatientForm from './NewPatientForm'
-import useTitle from '../../page-header/useTitle'
-import Patient from '../../model/Patient'
+import useTitle from '../../../page-header/useTitle'
+import Patient from '../../../model/Patient'
 import { createPatient } from '../patients-slice'
 
 const NewPatient = () => {
