+++ src/persons/patients/view/ViewPatient.tsx
@@ -3,11 +3,11 @@ import { useDispatch, useSelector } from 'react-redux'
 import { useParams, withRouter, Route, useHistory, useLocation } from 'react-router-dom'
 import { Panel, Spinner, TabsHeader, Tab } from '@hospitalrun/components'
 import { useTranslation } from 'react-i18next'
-import useTitle from '../../page-header/useTitle'
+import useTitle from '../../../page-header/useTitle'
 import { fetchPatient } from '../patient-slice'
-import { RootState } from '../../store'
+import { RootState } from '../../../store'
 import { getPatientFullName } from '../util/patient-name-util'
-import Patient from '../../model/Patient'
+import Patient from '../../../model/Patient'
 import GeneralInformation from './GeneralInformation'
 import RelatedPerson from '../related-persons/RelatedPersonTab'
